+++ lib/detector.dart
@@ -1,0 +1,93 @@
+import 'package:demo_app/detect_on_image.dart';
+import 'package:demo_app/detect_on_video.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_speed_dial/flutter_speed_dial.dart';
+import 'package:flutter_vision/flutter_vision.dart';
+
+enum Options { none, image, frame, vision }
+
+class Detector extends StatefulWidget {
+  const Detector({super.key});
+
+  @override
+  State<Detector> createState() => _DetectorState();
+}
+
+class _DetectorState extends State<Detector> {
+  late FlutterVision vision;
+  Options option = Options.none;
+  @override
+  void initState() {
+    super.initState();
+    vision = FlutterVision();
+  }
+
+  @override
+  void dispose() async {
+    super.dispose();
+
+    await vision.closeYoloModel();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: task(option),
+      floatingActionButton: SpeedDial(
+        //margin bottom
+        icon: Icons.menu, //icon on Floating action button
+        activeIcon: Icons.close, //icon when menu is expanded on button
+        backgroundColor: Colors.black12, //background color of button
+        foregroundColor: Colors.white, //font color, icon color in button
+        activeBackgroundColor:
+            Colors.deepPurpleAccent, //background color when menu is expanded
+        activeForegroundColor: Colors.white,
+        visible: true,
+        closeManually: false,
+        curve: Curves.bounceIn,
+        overlayColor: Colors.black,
+        overlayOpacity: 0.5,
+        buttonSize: const Size(56.0, 56.0),
+        children: [
+          SpeedDialChild(
+            //speed dial child
+            child: const Icon(Icons.video_call),
+            backgroundColor: Colors.red,
+            foregroundColor: Colors.white,
+            label: 'From Video',
+            labelStyle: const TextStyle(fontSize: 18.0),
+            onTap: () {
+              setState(() {
+                option = Options.frame;
+              });
+            },
+          ),
+          SpeedDialChild(
+            child: const Icon(Icons.camera),
+            backgroundColor: Colors.blue,
+            foregroundColor: Colors.white,
+            label: 'From Image',
+            labelStyle: const TextStyle(fontSize: 18.0),
+            onTap: () {
+              setState(() {
+                option = Options.image;
+              });
+            },
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget task(Options option) {
+    if (option == Options.frame) {
+      return DetectOnVideo(vision: vision);
+    }
+
+    if (option == Options.image) {
+      return DetectOnImage(vision: vision);
+    }
+
+    return const Center(child: Text("Choose Task"));
+  }
+}
