+++ test/widget_test.dart
@@ -5,17 +5,14 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
+import 'package:demo_app/detector.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:demo_app/main.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(MainApp(
-      cameras: [],
-    ));
+    await tester.pumpWidget(const Detector());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
